+++ contracts/name-minter/src/state.rs
@@ -6,9 +6,9 @@ use cw_storage_plus::Item;
 #[cw_serde]
 pub struct SudoParams {
     /// 3
-    pub min_name_length: u64,
+    pub min_name_length: u32,
     /// 63
-    pub max_name_length: u64,
+    pub max_name_length: u32,
     /// 100_000_000
     pub base_price: u128,
 }
